requirements.txt
amqp==2.4.2
argh==0.26.2
asn1crypto==0.24.0
audioread==2.1.5
autopep8==1.4.4
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
billiard==3.6.0.0
bleach==1.4.3
celery==4.3.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudinary==1.11.0
cryptography==2.1.4
Cython==0.29.3
decorator==4.2.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.11.11
django-advanced-filters==1.0.7.1
django-appconf==1.0.3
django-assets==0.12
django-auditlog==0.4.5
django-braces==1.4.0
django-compressor==2.2
django-easy-pdf==0.1.1
django-extensions==2.1.5
django-filter==1.1.0
django-filters==0.2.1
django-jsonfield==1.2.0
django-model-changes==0.15
django-model-changes-py3==0.15.1
django-money==0.14.4
django-nested-formset==0.1.4
django-postgres-extensions==0.9.3
django-qrcode==0.3
django-render-partial==0.3
django-reset-migrations==0.3.1
django-silk==3.0.2
django-suit==0.2.26
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
djorm-ext-pgarray==0.10
dnspython==1.15.0
docx2csv==0.1.0
easy-thumbnails==2.6
enum34==1.1.6
et-xmlfile==1.0.1

